Extract some common js functions out
